+++ examples/Benchmark/WithProgressBar/benchmark-w-reset.php
@@ -14,8 +14,6 @@ const ITERATIONS = 900000;
 
 require_once __DIR__ . '/../../../vendor/autoload.php';
 
-
-
 $benchmark = new BenchmarkSymfonyPB(ITERATIONS);
 $progressBar = $benchmark->getProgressBar();
 $progressBar->setBarWidth(60);
@@ -25,28 +23,16 @@ $progressBar->setFormat('[%bar%] %percent:3s%% %elapsed:6s%/%estimated:-6s%');
 $benchmark
     ->withComment('floatval()')
     ->addFunction('floatval', '3.5');
-
-$benchmark
-    ->withComment('intval()')
-    ->addFunction('intval', '3');
-
-$benchmark->run();
-$report = $benchmark->getReport();
-echo $report . PHP_EOL;
-
-$benchmark->reset();
 $benchmark
     ->withComment('(float)')
     ->addFunction(function () {
         return (float)'3.5';
     });
-
 $benchmark
     ->withComment('float "+"')
     ->addFunction(function () {
         return +'3.5';
     });
-
 
 $benchmark->run();
 $report = $benchmark->getReport();
@@ -65,7 +51,11 @@ $benchmark
         return +'3';
     });
 
+$benchmark
+    ->withComment('intval()')
+    ->addFunction('intval', '3');
 $benchmark->run();
 $report = $benchmark->getReport();
 echo $report . PHP_EOL;
+
 echo $benchmark->stat() . PHP_EOL;